lab5/pythonlabsem2/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[full]
python-dotenv
rarfile
pydantic
pydantic-settings